% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotters.R
\name{sparkline_area}
\alias{sparkline_area}
\title{Generate a sparkline area chart from a vector of numbers}
\usage{
sparkline_area(numbers)
}
\arguments{
\item{numbers}{Vector of numbers.}
}
\value{
Sparkline plot as a character vector of length 1
}
\description{
Generate a sparkline area chart from a vector of numbers
}
\examples{
test1 <- c(1, 2, 3, 4, 5, 6, 7, 8, 7, 6, 5, 4, 3, 2, 1)
test2 <- c(1.5, 0.5, 3.5, 2.5, 5.5, 4.5, 7.5, 6.5)
sparkline(test1)
sparkline(test2)
}
