% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotters.R
\name{sparkline_bars}
\alias{sparkline_bars}
\title{Generate a sparkline bar chart from a vector of numbers. This merely calls
on the \code{sparkline()} function with the \code{chars} arg set to "bars".}
\usage{
sparkline_bars(numbers)
}
\arguments{
\item{numbers}{Vector of numbers.}
}
\value{
Sparkline plot as a character vector of length 1
}
\description{
Generate a sparkline bar chart from a vector of numbers. This merely calls
on the \code{sparkline()} function with the \code{chars} arg set to "bars".
}
\examples{
test1 <- c(1, 2, 3, 4, 5, 6, 7, 8, 7, 6, 5, 4, 3, 2, 1)
test2 <- c(1.5, 0.5, 3.5, 2.5, 5.5, 4.5, 7.5, 6.5)
sparkline_bar(test1)
sparkline_bar(test2)
}
