% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotters.R, R/plotters_copy.R
\name{sparkline}
\alias{sparkline}
\title{Generate a sparkline from a vector of numbers}
\usage{
sparkline(numbers, return_range = F, symbols = c("bars", "lines"))

sparkline(numbers, return_range = F, symbols = c("bars", "lines"))
}
\arguments{
\item{numbers}{Vector of numbers.}

\item{return_range}{Logical. If \code{TRUE}, returns min and max values of
input vector along with sparkline plot.}

\item{symbols}{Symbols used to create sparkline. Either \code{'lines'} or
\code{'bars'}.}
}
\value{
Sparkline plot as a character vector of length 1. Optionally includes
  min and max values per value of \code{return_range}.

Sparkline plot as a character vector of length 1. Optionally includes
  min and max values per value of \code{return_range}.
}
\description{
Generate a sparkline from a vector of numbers

Generate a sparkline from a vector of numbers
}
\examples{
test1 <- c(1, 2, 3, 4, 5, 6, 7, 8, 7, 6, 5, 4, 3, 2, 1)
test2 <- c(1.5, 0.5, 3.5, 2.5, 5.5, 4.5, 7.5, 6.5)
sparkline(test1)
sparkline(test2)
test1 <- c(1, 2, 3, 4, 5, 6, 7, 8, 7, 6, 5, 4, 3, 2, 1)
test2 <- c(1.5, 0.5, 3.5, 2.5, 5.5, 4.5, 7.5, 6.5)
sparkline(test1)
sparkline(test2)
}
